Use float for updated() itself

+++ lib/Controller/ItemApiController.php
@@ -122,12 +122,12 @@ class ItemApiController extends ApiController
      *
      * @param int $type
      * @param int $id
-     * @param int $lastModified
+     * @param float $lastModified
      * @return array|JSONResponse
      *
      * @throws ServiceValidationException
      */
-    public function updated(int $type = 3, int $id = 0, int $lastModified = 0): array
+    public function updated(int $type = 3, int $id = 0, float $lastModified = 0): array
     {
         // needs to be turned into a microsecond timestamp to work properly
         if (strlen((string) $lastModified) <= 10) {
